feat: 회원가입 기능 구현
- 아이디, 이메일 중복 체크 구현
- 이메일 인증 구현
- 회원가입 insert 구현
- yml 암호화, mail, jasypt 설정
- security 기본 설정
- build.gradle 추가

FDORA-47

+++ src/main/java/com/farmdora/farmdoraauth/entity/Gender.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public enum Gender {
-    MALE(1), FEMALE(2);
+    MALE((byte)1), FEMALE((byte)2);
 
-    private final int value;
+    private final byte value;
 }
